Move region plotting to mpl class.
