+++ discretisedfield/plotting/__init__.py
@@ -1,3 +1,4 @@
 """Matplotlib based plotting."""
 from discretisedfield.plotting.mpl import Mpl
+from discretisedfield.plotting.mpl_region import MplRegion
 from discretisedfield.plotting.k3d import K3d
